Move ProjectionExprs into cse since it is now an implementation detail
